# config/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "operator",
        "version": "1.0",
        "description": "オペレーターサービス"
    },
    "servers": [
        {
            "url": "/operator"
        }
    ],
    "paths": {
...
        "/user/info": {
            "get": {
                "tags": [
                    "利用者情報"
                ],
                "description": "利用者情報 取得API",
                "parameters": [
                    {
                        "in": "header",
                        "name": "session",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "セッション情報",
                            "example": {
                                "sessionId": "a1da98a9a3f146b7418ddb5267bb241b0c603c61c8102f3e1e25f08aee375223",
                                "operatorId": 1,
                                "type": 3,
                                "loginId": "mng_member",
                                "name": "mng_member",
                                "auth": {
                                    "member": {
                                        "add": true,
                                        "update": true,
                                        "delete": true
                                    }
                                },
                                "lastLoginAt": "2020-01-01 00:00:00.000+0900",
                                "attributes": {},
                                "block": {
                                    "_value": 1000110,
                                    "_ver": 1
                                },
                                "actor": {
                                    "_value": 1000001,
                                    "_ver": 1
                                }
                            }
                        }
                    },
                    {
                        "in": "query",
                        "name": "pxrId",
                        "description": "PXR-ID",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "リクエストのオペレーターに紐づく利用者情報",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "利用者情報",
                                    "properties": {
                                        "pxrId": {
                                            "type": "string",
                                            "description": "PXR-ID",
                                            "example": "agent"
                                        },
                                        "userInfo": {
                                            "$ref": "#/components/schemas/userInfo"
                                        }
                                    }
                                },
                                "example": {
                                    "pxrId": "agent",
                                    "userInfo": {
                                        "_code": {
                                            "_value": 1000373,
                                            "_ver": 1
                                        },
                                        "'item-group'": [
                                            {
                                                "title": "氏名",
                                                "item": [
                                                    {
                                                        "title": "姓",
                                                        "type": {
                                                            "_value": 30019,
                                                            "_ver": 1
                                                        },
                                                        "content": "サンプル"
                                                    },
                                                    {
                                                        "title": "名",
                                                        "type": {
                                                            "_value": 30020,
                                                            "_ver": 1
                                                        },
                                                        "content": "太郎"
                                                    }
                                                ]
                                            },
                                            {
                                                "title": "性別",
                                                "item": [
                                                    {
                                                        "title": "性別",
                                                        "type": {
                                                            "_value": 30021,
                                                            "_ver": 1
                                                        },
                                                        "content": "男"
                                                    }
                                                ]
                                            },
                                            {
                                                "title": "生年",
                                                "item": [
                                                    {
                                                        "title": "生年",
                                                        "type": {
                                                            "_value": 1000372,
                                                            "_ver": 1
                                                        },
                                                        "content": 2000
                                                    }
                                                ]
                                            },
                                            {
                                                "title": "住所（行政区）",
                                                "item": [
                                                    {
                                                        "title": "住所（行政区）",
                                                        "type": {
                                                            "_value": 1000371,
                                                            "_ver": 1
                                                        },
                                                        "content": "東京都港区"
                                                    }
                                                ]
                                            },
                                            {
                                                "title": "連絡先電話番号",
                                                "item": [
                                                    {
                                                        "title": "連絡先電話番号",
                                                        "type": {
                                                            "_value": 30036,
                                                            "_ver": 1
                                                        },
                                                        "content": "080-1234-5678",
                                                        "changable-flag": true,
                                                        "require-sms-verification": true
                                                    }
                                                ]
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "利用者情報"
                ],
                "description": "利用者情報 登録API",
                "parameters": [
                    {
                        "in": "header",
                        "name": "session",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "セッション情報",
                            "example": {
                                "sessionId": "a1da98a9a3f146b7418ddb5267bb241b0c603c61c8102f3e1e25f08aee375223",
                                "operatorId": 1,
                                "type": 3,
                                "loginId": "mng_member",
                                "name": "mng_member",
                                "auth": {
                                    "member": {
                                        "add": true,
                                        "update": true,
                                        "delete": true
                                    }
                                },
                                "lastLoginAt": "2020-01-01 00:00:00.000+0900",
                                "attributes": {},
                                "block": {
                                    "_value": 1000110,
                                    "_ver": 1
                                },
                                "actor": {
                                    "_value": 1000001,
                                    "_ver": 1
                                }
                            }
                        }
                    }
                ],
                "requestBody": {
                    "description": "更新対象のオペレーター情報, 登録する利用者情報",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "利用者情報",
                                "properties": {
                                    "pxrId": {
                                        "type": "string",
                                        "description": "PXR-ID",
                                        "example": "agent"
                                    },
                                    "userInfo": {
                                        "$ref": "#/components/schemas/userInfo"
                                    }
                                }
                            },
                            "example": {
                                "pxrId": "agent",
                                "userInfo": {
                                    "_code": {
                                        "_value": 1000373,
                                        "_ver": 1
                                    },
                                    "'item-group'": [
                                        {
                                            "title": "氏名",
                                            "item": [
                                                {
                                                    "title": "姓",
                                                    "type": {
                                                        "_value": 30019,
                                                        "_ver": 1
                                                    },
                                                    "content": "サンプル"
                                                },
                                                {
                                                    "title": "名",
                                                    "type": {
                                                        "_value": 30020,
                                                        "_ver": 1
                                                    },
                                                    "content": "太郎"
                                                }
                                            ]
                                        },
                                        {
                                            "title": "性別",
                                            "item": [
                                                {
                                                    "title": "性別",
                                                    "type": {
                                                        "_value": 30021,
                                                        "_ver": 1
                                                    },
                                                    "content": "男"
                                                }
                                            ]
                                        },
                                        {
                                            "title": "生年",
                                            "item": [
                                                {
                                                    "title": "生年",
                                                    "type": {
                                                        "_value": 1000372,
                                                        "_ver": 1
                                                    },
                                                    "content": 2000
                                                }
                                            ]
                                        },
                                        {
                                            "title": "住所（行政区）",
                                            "item": [
                                                {
                                                    "title": "住所（行政区）",
                                                    "type": {
                                                        "_value": 1000371,
                                                        "_ver": 1
                                                    },
                                                    "content": "東京都港区"
                                                }
                                            ]
                                        },
                                        {
                                            "title": "連絡先電話番号",
                                            "item": [
                                                {
                                                    "title": "連絡先電話番号",
                                                    "type": {
                                                        "_value": 30036,
                                                        "_ver": 1
                                                    },
                                                    "content": "080-1234-5678",
                                                    "changable-flag": true,
                                                    "require-sms-verification": true
                                                }
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "登録したオペレーターの情報",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "pxrId": {
                                            "type": "string",
                                            "description": "PXR-ID",
                                            "example": "agent"
                                        }
                                    }
                                },
                                "example": {
                                    "pxrId": "agent"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "利用者情報"
                ],
                "description": "利用者情報 削除API",
                "parameters": [
                    {
                        "in": "header",
                        "name": "session",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "セッション情報",
                            "example": {
                                "sessionId": "a1da98a9a3f146b7418ddb5267bb241b0c603c61c8102f3e1e25f08aee375223",
                                "operatorId": 1,
                                "type": 3,
                                "loginId": "mng_member",
                                "name": "mng_member",
                                "auth": {
                                    "member": {
                                        "add": true,
                                        "update": true,
                                        "delete": true
                                    }
                                },
                                "lastLoginAt": "2020-01-01 00:00:00.000+0900",
                                "attributes": {},
                                "block": {
                                    "_value": 1000110,
                                    "_ver": 1
                                },
                                "actor": {
                                    "_value": 1000001,
                                    "_ver": 1
                                }
                            }
                        }
                    },
                    {
                        "in": "query",
                        "name": "pxrId",
                        "description": "PXR-ID",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "削除したオペレーターの情報",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "pxrId": {
                                            "type": "string",
                                            "description": "PXR-ID",
                                            "example": "agent"
                                        }
                                    }
                                },
                                "example": {
                                    "pxrId": "agent"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "userInfo": {
                "type": "object",
                "properties": {
                    "_code": {
                        "type": "object",
                        "description": "利用者情報 ひな形情報カタログ",
                        "properties": {
                            "_value": {
                                "type": "number",
                                "description": "カタログコード値",
                                "example": 1000373
                            },
                            "_ver": {
                                "type": "number",
                                "description": "カタログバージョン値",
                                "example": 1
                            }
                        }
                    },
                    "item-group": {
                        "type": "array",
                        "description": "利用者情報 各アイテム",
                        "items": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "description": "タイトル（アイテム名称）",
                                    "example": "氏名"
                                },
                                "item": {
                                    "type": "object",
                                    "description": "利用者 個別アイテム",
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "description": "タイトル（個別アイテム名称）",
                                            "example": "姓"
                                        },
                                        "type": {
                                            "type": "object",
                                            "description": "個別アイテム情報カタログ",
                                            "properties": {
                                                "_value": {
                                                    "type": "number",
                                                    "description": "個別アイテム カタログコード値",
                                                    "example": 30019
                                                },
                                                "_ver": {
                                                    "type": "number",
                                                    "description": "個別アイテム カタログバージョン値",
                                                    "example": 1
                                                }
                                            }
                                        },
                                        "content": {
                                            "type": "string",
                                            "description": "そのアイテムの値",
                                            "example": "太郎"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}